+++ src/main/java/io/github/astrapi69/data/identifiable/IdGenerator.java
@@ -26,8 +26,11 @@ package io.github.astrapi69.data.identifiable;
 
 /**
  * The interface {@link IdGenerator} provides a method for get the next id
+ *
+ * @param <T>
+ *            the generic type of the id
  */
-public interface IdGenerator
+public interface IdGenerator<T>
 {
 
 	/**
@@ -35,6 +38,6 @@ public interface IdGenerator
 	 * 
 	 * @return the next id
 	 */
-	int getNextId();
+	T getNextId();
 
 }